Ластецкая

+++ src/main/java/net/thumbtack/onlineshop/service/impl/CategoryServiceImpl.java
@@ -15,10 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
@@ -42,6 +39,9 @@ public class CategoryServiceImpl implements CategoryService {
         SessionServiceImpl sessionService = getSessionService(administratorRepository, clientRepository, sessionRepository);
         if (sessionService.isAdmin(sessionService.getLogin(cookie))) {
             Category category = new Category(request.getName(), request.getIdParentCategory());
+            if (category.getIdParentCategory() == null) {
+                category.setIdParentCategory(0);
+            }
             category = categoryRepository.save(category);
             return new AddCategoryResponse(category.getId(), category.getName(), category.getIdParentCategory(),
                     getParentName(category.getIdParentCategory()));
@@ -53,24 +53,29 @@ public class CategoryServiceImpl implements CategoryService {
     public AddCategoryResponse getCategoryById(String cookie, Integer id) throws OnlineShopExceptionOld {
         SessionServiceImpl sessionService = getSessionService(administratorRepository, clientRepository, sessionRepository);
         if (sessionService.isAdmin(sessionService.getLogin(cookie))) {
-            Category category = categoryRepository.findById(id).get();
-            return new AddCategoryResponse(category.getId(), category.getName(), category.getIdParentCategory(),
-                    getParentName(category.getIdParentCategory()));
+            if (categoryRepository.existsById(id)) {
+                Category category = categoryRepository.findById(id).get();
+                return new AddCategoryResponse(category.getId(), category.getName(), category.getIdParentCategory(),
+                        getParentName(category.getIdParentCategory()));
+            }
         } else {
-            return null;
+            throw new OnlineShopExceptionOld("Данное действие доступно только для администратора");
         }
+        return null;
     }
 
     @Override
     public AddCategoryResponse editCategory(String cookie, EditCategoryRequest request, Integer id) throws OnlineShopExceptionOld {
         SessionServiceImpl sessionService = getSessionService(administratorRepository, clientRepository, sessionRepository);
-        if(request.getIdParent() == null && request.getName() == null) {
-            throw new OnlineShopExceptionOld();
+        if (request.getIdParent() == null && request.getName() == null) {
+            throw new OnlineShopExceptionOld("Оба поля не могут быть пустыми");
         }
         if (sessionService.isAdmin(sessionService.getLogin(cookie))) {
             Category category = categoryRepository.findById(id).get();
             if ((category.getIdParentCategory() == 0 || category.getIdParentCategory() == null) && (request.getIdParent() != 0)) {
-                throw new OnlineShopExceptionOld();
+                throw new OnlineShopExceptionOld("Категория не может стать подкатегорией");
+            } else if (category.getIdParentCategory() != 0 && (request.getIdParent() == 0 || request.getIdParent() == null)) {
+                throw new OnlineShopExceptionOld("Подкатегория не может стать категорией");
             } else {
                 if (StringUtils.isEmpty(request.getName())) {
                     category.setIdParentCategory(request.getIdParent());
@@ -87,17 +92,22 @@ public class CategoryServiceImpl implements CategoryService {
                 return new AddCategoryResponse(category.getId(), category.getName(), category.getIdParentCategory(),
                         getParentName(category.getIdParentCategory()));
             }
+        } else {
+            throw new OnlineShopExceptionOld("Данное действие доступно только для администратора");
         }
-        return null;
     }
 
     @Override
     public void deleteCategoryById(String cookie, Integer id) throws OnlineShopExceptionOld {
         SessionServiceImpl sessionService = getSessionService(administratorRepository, clientRepository, sessionRepository);
         if (sessionService.isAdmin(sessionService.getLogin(cookie))) {
-            categoryRepository.deleteById(id);
+            if (categoryRepository.existsById(id)) {
+                categoryRepository.deleteById(id);
+            } else {
+                throw new OnlineShopExceptionOld("Такой категории не существует");
+            }
         } else {
-            throw new OnlineShopExceptionOld();
+            throw new OnlineShopExceptionOld("Данное действие доступно только для администратора");
         }
     }
 
@@ -106,10 +116,10 @@ public class CategoryServiceImpl implements CategoryService {
         Iterable<Category> categories = categoryRepository.findAllByIdParentCategoryEqualsOrderByNameAsc(0);
         Iterable<Category> subCategories = categoryRepository.findAllByIdParentCategoryGreaterThanOrderByNameAsc(0);
         List<AddCategoryResponse> responseList = new ArrayList<>();
-        for(Category category : categories) {
+        for (Category category : categories) {
             responseList.add(new AddCategoryResponse(category.getId(), category.getName(),
                     category.getIdParentCategory()));
-            for(Category subCategory : subCategories) {
+            for (Category subCategory : subCategories) {
                 if (subCategory.getIdParentCategory().equals(category.getId())) {
                     responseList.add(new AddCategoryResponse(subCategory.getId(), subCategory.getName(),
                             subCategory.getIdParentCategory()));
